+++ src/lib.rs
@@ -37,3 +37,30 @@ mod search;
 mod sort_by;
 
 pub use {client::Client, fields::Fields, page::Page, search::Search, sort_by::SortBy};
+
+/// Crate-specific error enum.
+#[derive(Debug, thiserror::Error)]
+pub enum Error {
+    /// A boxed error.
+    ///
+    /// Used to capture generic errors from [tokio_postgres::types::FromSql].
+    #[error(transparent)]
+    Boxed(#[from] Box<dyn std::error::Error + Sync + Send>),
+
+    /// [serde_json::Error]
+    #[error(transparent)]
+    SerdeJson(#[from] serde_json::Error),
+
+    /// [tokio_postgres::Error]
+    #[error(transparent)]
+    TokioPostgres(#[from] tokio_postgres::Error),
+
+    /// An unknown error.
+    ///
+    /// Used when [tokio_postgres::types::FromSql] doesn't have a source.
+    #[error("unknown error")]
+    Unknown,
+}
+
+/// Crate-specific result type.
+pub type Result<T> = std::result::Result<T, Error>;