+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Changed
+
+- Make `Error` and `Result` publicly visible
+
 ## [0.0.1] - 2023-01-07
 
 Initial release